samples-master/best_practices/contrib/part_detection/inference/requirements.txt
torch==1.10.1
torchvision==0.11.2
onnx==1.12.0
onnxruntime>=1.8.0
pycocotools>=2.0
numpy>=1.18.5
opencv-python>=4.1.1
Pillow>=7.1.2
scipy>=1.4.1
tqdm>=4.41.0
seaborn>=0.11.0
pyyaml>=5.3.1